refactor: move generated client instantiation out of base class
